Working

+++ src/closed_loop/src/splitter.cpp
@@ -5,7 +5,7 @@
 
 #include "rclcpp/rclcpp.hpp"
 #include "asdfr_interfaces/msg/point2.hpp"
-#include "std_msgs/msg/float32.hpp"
+#include "std_msgs/msg/float64.hpp"
 
 using std::placeholders::_1;
 using namespace std::chrono_literals;
@@ -14,13 +14,13 @@ class Splitter : public rclcpp::Node
 {
 private:
     rclcpp::Subscription<asdfr_interfaces::msg::Point2>::SharedPtr input_sub_;
-    rclcpp::Publisher<std_msgs::msg::Float32>::SharedPtr output_x_pub_;
-    rclcpp::Publisher<std_msgs::msg::Float32>::SharedPtr output_y_pub_;
+    rclcpp::Publisher<std_msgs::msg::Float64>::SharedPtr output_x_pub_;
+    rclcpp::Publisher<std_msgs::msg::Float64>::SharedPtr output_y_pub_;
 
     void on_input(const asdfr_interfaces::msg::Point2 msg) {
 
-        std_msgs::msg::Float32 x;
-        std_msgs::msg::Float32 y;
+        std_msgs::msg::Float64 x;
+        std_msgs::msg::Float64 y;
 
         x.data = msg.x;
         y.data = msg.y;
@@ -37,8 +37,8 @@ public:
         input_sub_ = this->create_subscription<asdfr_interfaces::msg::Point2>(
         "input", 1 , std::bind(&Splitter::on_input, this, _1));
 
-        output_x_pub_ = this->create_publisher<std_msgs::msg::Float32>("output_x", 1);
-        output_y_pub_ = this->create_publisher<std_msgs::msg::Float32>("output_y", 1);
+        output_x_pub_ = this->create_publisher<std_msgs::msg::Float64>("output_x", 1);
+        output_y_pub_ = this->create_publisher<std_msgs::msg::Float64>("output_y", 1);
     }
 
 };
